refactor: 大规模代码重构与依赖清理
- 迁移并重构大量包路径，统一代码组织规范
- 删除废弃的老旧模块类：VmaMemoryBudget、LayerBatchMerger等十余种过时实现
- 新增配置系统基础接口与枚举类型，完善配置框架
- 修复包导入错误，统一依赖引用路径
- 优化渲染管线相关代码，移除冗余导入与重复声明
- 更新CI构建仓库源配置，适配国内加速镜像
- 重构可见性结果判断逻辑，修复数组越界隐患

+++ common/src/main/java/com/ranecc/renderium/feature/culling/intercept/CullingAdapter.java
@@ -7,7 +7,7 @@ import com.ranecc.renderium.feature.culling.optimization.NeighborFaceCuller;
 import com.ranecc.renderium.feature.culling.optimization.HeightmapOcclusionCuller;
 import com.ranecc.renderium.feature.culling.optimization.AsyncComputeCuller;
 
-import com.ranecc.renderium.None;
+import com.ranecc.renderium.feature.culling.intercept.InterceptionCullingContext;
 
 import java.util.BitSet;
 import java.util.LinkedHashMap;
@@ -492,7 +492,7 @@ public final class CullingAdapter {
             } else {
                 // ======== 无方块状态数据：基于历史统计返回估算结果 ========
                 // 邻居面剔除通常是离线/预计算的，此处使用历史剔除率进行估算
-                float historicalCullRate = culler.getCullRate();
+                float historicalCullRate = (float) culler.getCullRate();
                 int estimatedCulledCount = (int) (objectCount * historicalCullRate);
 
                 // 创建估算的可见性掩码（随机分布，仅用于统计）
